fixed warning in TMCParticle.cxx:54: warning: passing Double_t for argument 1 of void

+++ mcroot/TMCParticle.cxx
@@ -51,7 +51,7 @@ TMCParticle::TMCParticle(TMCFastHepParticle &heppart,TMCFastOfflineTrack &offtrk
   SetX(heppart.GetVx()) ; 
   SetY(heppart.GetVy()) ;
   SetZ(heppart.GetVz()) ;
-  SetCharge(offtrk.GetQ(track));
+  SetCharge((int)offtrk.GetQ(track));
   SetIdHep(heppart.GetIdhep()) ;
   SetHepIndex(offtrk.GetHep(track));
   SetMaker(MADE_FROM_OFFTRK) ;
